# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MailDay Matching API
servers:
  - url: /api
...
paths:
...
  /children:
    get:
      operationId: listChildren
      tags: [children]
      summary: List all children
      parameters:
        - name: match_status
          in: query
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Children list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChildWithParent"
    post:
      operationId: createChild
      tags: [children]
      summary: Create a new child record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChildInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
  /children/{id}:
    get:
      operationId: getChild
      tags: [children]
      summary: Get a single child
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Child record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChildWithParent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateChild
      tags: [children]
      summary: Update a child record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChildUpdate"
      responses:
        "200":
          description: Updated child
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Parent:
      type: object
      required: [id, first_name, last_name, email]
      properties:
        id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        mailing_address:
          type: ["string", "null"]
        address_type:
          type: ["string", "null"]
        shopify_customer_id:
          type: ["string", "null"]
        recharge_subscription_id:
          type: ["string", "null"]
        membership_tier:
          type: ["string", "null"]
        billing_type:
          type: ["string", "null"]
        subscription_status:
          type: ["string", "null"]
        join_date:
          type: ["string", "null"]
        referral_source:
          type: ["string", "null"]
        community_status:
          type: ["string", "null"]
        give_a_key_recipient:
          type: ["boolean", "null"]
        last_email_open_date:
          type: ["string", "null"]
        at_risk:
          type: ["boolean", "null"]
        internal_notes:
          type: ["string", "null"]
        created_at:
          type: string
...
    Child:
      type: object
      required: [id, parent_id, child_first_name, age, tier, match_status]
      properties:
        id:
          type: string
        parent_id:
          type: string
        child_first_name:
          type: string
        age:
          type: integer
        tier:
          type: string
          description: Core / Minis / Homeschool Core / Homeschool Minis
        interests:
          type: array
          items:
            type: string
        homeschool_edition:
          type: ["boolean", "null"]
        homeschool_tier:
          type: ["string", "null"]
        homeschool_approach:
          type: ["string", "null"]
        match_status:
          type: string
          description: Unmatched / Matched / Rematch Requested / Rematched / Paused / Cancelled
        rematch_count:
          type: ["integer", "null"]
        match_guarantee_start_date:
          type: ["string", "null"]
        billing_paused:
          type: ["boolean", "null"]
        safety_flag:
          type: ["boolean", "null"]
        internal_notes:
          type: ["string", "null"]
        created_date:
          type: ["string", "null"]
        created_at:
          type: string
    ChildInput:
      type: object
      required: [parent_id, child_first_name, age, tier]
      properties:
        parent_id:
          type: string
        child_first_name:
          type: string
        age:
          type: integer
        tier:
          type: string
        interests:
          type: array
          items:
            type: string
        homeschool_edition:
          type: boolean
        homeschool_tier:
          type: string
        homeschool_approach:
          type: string
        safety_flag:
          type: boolean
        internal_notes:
          type: string
    ChildUpdate:
      type: object
      properties:
        child_first_name:
          type: string
        age:
          type: integer
        tier:
          type: string
        interests:
          type: array
          items:
            type: string
        homeschool_edition:
          type: boolean
        homeschool_tier:
          type: string
        homeschool_approach:
          type: string
        match_status:
          type: string
        rematch_count:
          type: integer
        match_guarantee_start_date:
          type: string
        billing_paused:
          type: boolean
        safety_flag:
          type: boolean
        internal_notes:
          type: string
    ChildWithParent:
      allOf:
        - $ref: "#/components/schemas/Child"
        - type: object
          required: [parent]
          properties:
            parent:
              $ref: "#/components/schemas/Parent"
            days_waiting:
              type: ["integer", "null"]
            guarantee_status:
              type: ["string", "null"]
              description: ok / warning / urgent
...